Fix tests for recent extension versions
Recent versions of the Tideways extension enabled Tracepoints in the CLI by
default. This leads to the extension sending the tracepoint request as the very
first thing that is sent to the daemon socket.

This leads to a Catch-22 when the daemon socket is bound to the same process:
We need to accept the tracepoint request to satisfy it and unblock the process,
but we can't accept it, since we are blocked on the response.

Furthermore the Tideways extension will send PHPInfo payloads even when
manually handling the lifecycle with `Profiler::start()` and
`Profiler::stop()`. This PHPInfo payload will need to be ignored as well.

Fix the tests by running the daemon as a background process. This background
daemon will ignore everything, except the trace payload and will then exit.