Replace dependency on `laravel/octane` by `illuminate/http`
While the middleware is intended to be used with Laravel Octane, it actually is
completely agnostic to how Laravel is executed. It does not use any
Octane-specific functionality.

Drop the dependency to avoid keeping track of Octane updates. This also
implicitly cleans up quite a few dependencies from composer.lock.